Program to seed Postgres
A program that leverages the random data generator to seed the postgres database.

Not fully decoupled yet as fields is an object defined *in* this document, instead of leveraging the Sequelize models.

Other enhancements would be to promisify the batches so that they can run sequentially. Currently, if you try to run them all at once, the connections trip over one another and all fail.