Refactor of fetchBudgetData for simplicity
Removed unnecessary comments.
Consolidated the function between its conditionals and a separate axios call.
Technically adding a second action of the fetchBudgetData function by first checking if it's a valid URL *then* making the AJAX call.
This feels like a reasonable consolidation and removes the extra logic living in the componentDidMount, which is now cleared up for future additions.