BudgetCategories can now be added through an add a category form
The form which continues to toggle on/off will pass the values from the form through to the app which will then process the records and add them to the database.
On a response, we will *refetch* the entire budget, giving a "quick flash" for a re-render, but it is more stable than adding a row on our own *without* confirmation that the line was actually committed.